refactor: extract probabilistic light client core
