+++ cosmos/cardano-probabilistic-light-client-core/go.mod
@@ -1,0 +1,21 @@
+module github.com/cardano-foundation/cardano-ibc-incubator/cosmos/cardano-probabilistic-light-client-core
+
+go 1.24.0
+
+require (
+	github.com/blinklabs-io/gouroboros v0.89.1
+	github.com/cosmos/ics23/go v0.11.0
+	github.com/fxamacker/cbor/v2 v2.7.0
+)
+
+require (
+	filippo.io/edwards25519 v1.1.0 // indirect
+	github.com/cosmos/gogoproto v1.7.0 // indirect
+	github.com/google/go-cmp v0.7.0 // indirect
+	github.com/jinzhu/copier v0.4.0 // indirect
+	github.com/utxorpc/go-codegen v0.5.1 // indirect
+	github.com/x448/float16 v0.8.4 // indirect
+	golang.org/x/crypto v0.46.0 // indirect
+	golang.org/x/sys v0.39.0 // indirect
+	google.golang.org/protobuf v1.36.10 // indirect
+)
