+++ cosmos/cardano-probabilistic-light-client-v10/internal/cardanodatum/types.go
@@ -1,199 +1,27 @@
 package cardanodatum
 
-import "github.com/blinklabs-io/gouroboros/ledger"
+import core "github.com/cardano-foundation/cardano-ibc-incubator/cosmos/cardano-probabilistic-light-client-core/cardanodatum"
 
-// This package contains CBOR/Plutus datum shapes used by the Cardano chain.
-//
-// The Mithril IBC light client needs these types to semantically compare values committed by
-// Cardano (which are CBOR-serialized datums) with the protobuf-encoded values used by ibc-go.
-//
-// It is intentionally *not* an IBC light client implementation: it is a small set of decoding
-// helpers that support the Mithril client's membership/non-membership verification logic.
-
-type SPOState struct {
-	IsRegisCert bool
-	PoolId      string
-	PoolVrf     string
-	BlockNo     uint64
-	TxIndex     uint64
-}
-
-type VerifyBlockOutput struct {
-	_             struct{} `cbor:",toarray"`
-	Flag          int
-	IsValid       bool
-	VrfKHexString string
-	BlockNo       string
-	SlotNo        string
-}
-
-type ExtractBlockOutput struct {
-	_            struct{} `cbor:",toarray"`
-	Flag         int
-	Outputs      []ledger.UTXOOutput
-	RegisCerts   []ledger.RegisCert
-	DeRegisCerts []ledger.DeRegisCert
-}
-
-//////////////////////////////////////////////////////
-// Struct types saved on the Cardano side (CBOR / PlutusData)
-
-// Client and consensus state
-type RootHashInDatum struct {
-	_    struct{} `cbor:",toarray"`
-	Hash []byte
-}
-
-type ConsensusStateDatum struct {
-	_                  struct{} `cbor:",toarray"`
-	Timestamp          uint64
-	NextValidatorsHash []byte
-	Root               RootHashInDatum
-}
-
-type TrustLevelDatum struct {
-	_           struct{} `cbor:",toarray"`
-	Numerator   uint64
-	Denominator uint64
-}
-
-type HeightDatum struct {
-	_              struct{} `cbor:",toarray"`
-	RevisionNumber uint64
-	RevisionHeight uint64
-}
-
-type LeafSpecDatum struct {
-	_            struct{} `cbor:",toarray"`
-	Hash         int32
-	PrehashKey   int32
-	PrehashValue int32
-	Length       int32
-	Prefix       []byte
-}
-
-type InnerSpecDatum struct {
-	_               struct{} `cbor:",toarray"`
-	ChildOrder      []int32
-	ChildSize       int32
-	MinPrefixLength int32
-	MaxPrefixLength int32
-	EmptyChild      []byte
-	Hash            int32
-}
-
-type ProofSpecsDatum struct {
-	_         struct{} `cbor:",toarray"`
-	LeafSpec  LeafSpecDatum
-	InnerSpec InnerSpecDatum
-	MaxDepth  int32
-	MinDepth  int32
-	// (PrehashKeyBeforeComparison.(cbor.Tag)).Number => False: 121, True: 122
-	PrehashKeyBeforeComparison interface{}
-}
-
-type ClientStateDatum struct {
-	_               struct{} `cbor:",toarray"`
-	ChainId         []byte
-	TrustLevel      TrustLevelDatum
-	TrustingPeriod  uint64
-	UnbondingPeriod uint64
-	MaxClockDrift   uint64
-	FrozenHeight    HeightDatum
-	LatestHeight    HeightDatum
-	ProofSpecs      []ProofSpecsDatum
-}
-
-type TokenDatum struct {
-	_        struct{} `cbor:",toarray"`
-	PolicyId []byte
-	Name     []byte
-}
-
-type ClientDatumState struct {
-	_               struct{} `cbor:",toarray"`
-	ClientState     ClientStateDatum
-	ConsensusStates map[HeightDatum]ConsensusStateDatum
-}
-
-type ClientDatum struct {
-	_     struct{} `cbor:",toarray"`
-	State ClientDatumState
-	Token TokenDatum
-}
-
-// Connection State
-type ConnectionDatum struct {
-	_     struct{} `cbor:",toarray"`
-	State ConnectionEndDatum
-	Token TokenDatum
-}
-
-type ConnectionEndDatum struct {
-	_        struct{} `cbor:",toarray"`
-	ClientId []byte
-	Versions []VersionDatum
-	// Little hack with this kind of Enum
-	// (State.(cbor.Tag)).Number => UNINITIALIZED: 121, INIT: 122, TRYOPEN: 123, OPEN: 124
-	State        interface{}
-	Counterparty CounterpartyDatum
-	DelayPeriod  uint64
-}
-
-type VersionDatum struct {
-	_          struct{} `cbor:",toarray"`
-	Identifier []byte
-	Features   [][]byte
-}
-
-type CounterpartyDatum struct {
-	_            struct{} `cbor:",toarray"`
-	ClientId     []byte
-	ConnectionId []byte
-	Prefix       MerklePrefixDatum
-}
-
-type MerklePrefixDatum struct {
-	_         struct{} `cbor:",toarray"`
-	KeyPrefix []byte
-}
-
-type ChannelCounterpartyDatum struct {
-	_         struct{} `cbor:",toarray"`
-	PortId    []byte
-	ChannelId []byte
-}
-
-type ChannelDatum struct {
-	_ struct{} `cbor:",toarray"`
-	// Little hack with this kind of Enum
-	// (State.(cbor.Tag)).Number => UNINITIALIZED: 121, INIT: 122, TRYOPEN: 123, OPEN: 124, CLOSED: 125
-	State interface{}
-	// Little hack with this kind of Enum
-	// (Ordering.(cbor.Tag)).Number => None: 121, Unordered: 122, Ordered: 123
-	Ordering       interface{}
-	Counterparty   ChannelCounterpartyDatum
-	ConnectionHops [][]byte
-	Version        []byte
-}
-
-type ChannelDatumState struct {
-	_                     struct{} `cbor:",toarray"`
-	Channel               ChannelDatum
-	NextSequenceSend      uint64
-	NextSequenceRecv      uint64
-	NextSequenceAck       uint64
-	PacketCommitment      map[uint64][]byte
-	PacketReceipt         map[uint64][]byte
-	PacketAcknowledgement map[uint64][]byte
-}
-
-type ChannelDatumWithPort struct {
-	_      struct{} `cbor:",toarray"`
-	State  ChannelDatumState
-	PortId []byte
-	Token  TokenDatum
-}
-
-//
-//////////////////////////////////////////////////////
+type SPOState = core.SPOState
+type VerifyBlockOutput = core.VerifyBlockOutput
+type ExtractBlockOutput = core.ExtractBlockOutput
+type RootHashInDatum = core.RootHashInDatum
+type ConsensusStateDatum core.ConsensusStateDatum
+type TrustLevelDatum = core.TrustLevelDatum
+type HeightDatum = core.HeightDatum
+type LeafSpecDatum = core.LeafSpecDatum
+type InnerSpecDatum = core.InnerSpecDatum
+type ProofSpecsDatum = core.ProofSpecsDatum
+type ClientStateDatum core.ClientStateDatum
+type TokenDatum = core.TokenDatum
+type ClientDatumState = core.ClientDatumState
+type ClientDatum = core.ClientDatum
+type ConnectionDatum = core.ConnectionDatum
+type ConnectionEndDatum core.ConnectionEndDatum
+type VersionDatum = core.VersionDatum
+type CounterpartyDatum = core.CounterpartyDatum
+type MerklePrefixDatum = core.MerklePrefixDatum
+type ChannelCounterpartyDatum = core.ChannelCounterpartyDatum
+type ChannelDatum core.ChannelDatum
+type ChannelDatumState = core.ChannelDatumState
+type ChannelDatumWithPort = core.ChannelDatumWithPort
