+++ cosmos/cardano-probabilistic-light-client-v8/host_state_commitment.go
@@ -1,11 +1,6 @@
 package probabilistic
 
-import (
-	"fmt"
-	"strings"
-
-	"github.com/blinklabs-io/gouroboros/ledger"
-)
+import probabilisticcore "github.com/cardano-foundation/cardano-ibc-incubator/cosmos/cardano-probabilistic-light-client-core"
 
 func (cs ClientState) ExtractIbcStateRootFromHostStateTx(header *ProbabilisticHeader) ([]byte, error) {
 	txBodyCbor, err := extractHostStateTxBodyCborFromAnchorBlock(header)
@@ -28,64 +23,13 @@ func extractIbcStateRootFromTransactionBody(
 	hostStateNftPolicyId []byte,
 	hostStateNftTokenName []byte,
 ) ([]byte, error) {
-	if txHash == "" {
-		return nil, fmt.Errorf("missing HostState transaction hash in header")
-	}
-	if len(txBodyCbor) == 0 {
-		return nil, fmt.Errorf("missing HostState transaction body CBOR in header")
-	}
-
-	txBody, err := decodeTransactionBody(txBodyCbor)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode HostState tx body: %w", err)
-	}
-	if strings.ToLower(txBody.Hash()) != strings.ToLower(txHash) {
-		return nil, fmt.Errorf("HostState tx body hash mismatch")
-	}
-
-	outputs := txBody.Outputs()
-	idx := int(outputIndex)
-	if idx < 0 || idx >= len(outputs) {
-		return nil, fmt.Errorf("HostState output index out of range")
-	}
-
-	out := outputs[idx]
-	if len(hostStateNftPolicyId) > 0 && len(hostStateNftTokenName) > 0 {
-		assets := out.Assets()
-		if assets == nil {
-			return nil, fmt.Errorf("HostState output has no multi-assets")
-		}
-		policy := ledger.NewBlake2b224(hostStateNftPolicyId)
-		if assets.Asset(policy, hostStateNftTokenName) != 1 {
-			return nil, fmt.Errorf("HostState output does not contain the expected HostState NFT")
-		}
-	}
-
-	datum := out.Datum()
-	if datum == nil {
-		return nil, fmt.Errorf("HostState output has no inline datum")
-	}
-	return ExtractIbcStateRootFromHostStateDatum(datum.Cbor(), hostStateNftPolicyId)
+	return probabilisticcore.ExtractIbcStateRootFromTransactionBody(
+		txBodyCbor,
+		txHash,
+		outputIndex,
+		hostStateNftPolicyId,
+		hostStateNftTokenName,
+	)
 }
 
-func decodeTransactionBody(data []byte) (ledger.TransactionBody, error) {
-	if body, err := ledger.NewConwayTransactionBodyFromCbor(data); err == nil {
-		return body, nil
-	}
-	if body, err := ledger.NewBabbageTransactionBodyFromCbor(data); err == nil {
-		return body, nil
-	}
-	if body, err := ledger.NewAlonzoTransactionBodyFromCbor(data); err == nil {
-		return body, nil
-	}
-	if body, err := ledger.NewMaryTransactionBodyFromCbor(data); err == nil {
-		return body, nil
-	}
-	if body, err := ledger.NewAllegraTransactionBodyFromCbor(data); err == nil {
-		return body, nil
-	}
-	if body, err := ledger.NewShelleyTransactionBodyFromCbor(data); err == nil {
-		return body, nil
-	}
-	return nil, fmt.Errorf("unsupported transaction body CBOR")
-}
+var decodeTransactionBody = probabilisticcore.DecodeTransactionBody
